Introduce timestamps
The chrono crate was added to the project in order to indicate
the time and date of relevant events during the lifetime of the
application. Now the start and end of the simulation, as well as
the transmission of messages to the Cloud, are timestamped.

This is an important step towards the incorporation of a logging
mechanism.